FEAT: Move tests to own files (and own npm scripts)
