+++ src/checks/webcoach/index.js
@@ -1,0 +1,19 @@
+// const URLS = require('../../config/urls')
+/*
+// ********************************************************************
+// Webcoach Scores
+// ********************************************************************
+const webcoachCheck = require('./src/checks/webcoach/check')
+
+for (let site in URLS) {
+  for (let type in URLS[site]) {
+    const url = URLS[site][type]
+    webcoachCheck.run(site, type, url, false).then(() => {
+      console.log(`Finished webcoach for ${url} on Desktop`)
+    })
+    webcoachCheck.run(site, type, url, true).then(() => {
+      console.log(`Finished webcoach for ${url} on Mobile`)
+    })
+  }
+}
+*/
